Extract kotlinpoet-only parts to new API package for better separation
