+++ kotlin/codegen/src/main/java/com/squareup/moshi/kotlin/codegen/api/TargetParameter.kt
@@ -13,23 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.squareup.moshi.kotlin.codegen
+package com.squareup.moshi.kotlin.codegen.api
 
-import com.squareup.kotlinpoet.KModifier
-import com.squareup.kotlinpoet.TypeName
-import com.squareup.kotlinpoet.TypeVariableName
+import com.squareup.kotlinpoet.AnnotationSpec
 
-/** A user type that should be decoded and encoded by generated code. */
-internal data class TargetType(
-  val typeName: TypeName,
-  val constructor: TargetConstructor,
-  val properties: Map<String, TargetProperty>,
-  val typeVariables: List<TypeVariableName>,
-  val isDataClass: Boolean,
-  val visibility: KModifier
-) {
-
-  init {
-    visibility.checkIsVisibility()
-  }
-}
+/** A parameter in user code that should be populated by generated code. */
+internal data class TargetParameter(
+  val name: String,
+  val index: Int,
+  val hasDefault: Boolean,
+  val jsonName: String = name,
+  val qualifiers: Set<AnnotationSpec>? = null
+)
